requirements.txt
pymongo
fastapi
uvicorn
pyyaml
passlib[bcrypt]